Fixed e2e test when navigating to ended hunt details
This wasn't as easy as I wish it was.  Needed a getter method which sends its own api get request to get the started hunt id.  This is the downside of using the access code for url navigation in started hunts (Which I was not for in iteration 2), which then gets turned to 1 after ending.  Should have been always using the id.